api/requirements.txt
faiss-cpu==1.7.4
fastapi==0.115.4
gensim==4.3.3
nltk==3.9.1
numpy==1.26.4
pandas==2.2.3
pydantic==1.10.18
torch==2.5.1
uvicorn>=0.21.0
pyarrow==14.0.1
psycopg2-binary==2.9.10
python-multipart==0.0.6
transformers==4.36.2
torchvision==0.20.1
Pillow==10.2.0
accelerate==0.26.1